Extract shared BaseConversationServiceInterface to a dedicated module, add GenerativeAnswerService class, and implement a new `generative_answer` method in the client to handle AnswerBasedOnTheContextModel payloads.
